refactor(graph): extract GraphCanvas.vue, GraphPanel becomes composer (Closes #34)
Letzte Etappe der GraphPanel-Zerlegung. D3-Renderlogik, SVG-Container,
Selektions-State (selectedItem, expandedSelfLoops, showEdgeLabels),
Loading/Empty-States, Edge-Labels-Toggle, Hint- und Detail-Panel-
Einbindung sowie alle vier Export-Funktionen wandern in
GraphCanvas.vue. GraphPanel.vue bleibt als reine Kompositionsdatei
zurück (98 Zeilen, −90 % gegenüber dem Ausgangsstand 933 Zeilen).

Toolbar-Click-Events werden via canvasRef weitergereicht: GraphCanvas
exposed downloadGraphml, downloadSvg, downloadPng, printPdf via
defineExpose; GraphPanel hält den ref und ruft beim entsprechenden
Toolbar-Event die passende Methode auf.

State-Verteilung:
- GraphPanel: selectedRound, showSimulationFinishedHint, wasSimulating,
  displayedGraphData (round-gefilterte Computed), entityTypes, maxRound
- GraphCanvas: selectedItem, expandedSelfLoops, showEdgeLabels,
  currentSimulation, linkLabelsRef, linkLabelBgRef

Akzeptanz Issue #34:
- GraphPanel.vue ist primär Kompositionsdatei ✓
- D3-Rendering nicht mit Detail/Hint-Logik vermischt ✓
- Zielstruktur (GraphCanvas/Toolbar/Hints/Legend/DetailPanel) vollständig ✓

D3-Logik bleibt vorerst in der Komponente — Composable-Extraktion ist
Issue #35 (EPIC-04-ST-02), eigene Story.

GraphPanel.vue: 736 → 98 Zeilen.
GraphCanvas.vue: +641 Zeilen.

Verifikation: npm run check 5/5 grün
- backend: 488 passed, 2 skipped
- frontend: 11 vitest passed, eslint clean, vite build (734 modules)

Closes #34
