+++ CHANGELOG.md
@@ -9,10 +9,11 @@ Format angelehnt an [Keep a Changelog](https://keepachangelog.com/de/1.1.0/), Ve
 
 - **`frontend/src/components/graph/GraphHints.vue`** — neue rein präsentationale Komponente für Building-/Simulating-Hint und Simulation-Finished-Hint. Erste Etappe von Issue #34 (EPIC-04-ST-01: GraphPanel zerlegen). Props: `currentPhase`, `isSimulating`, `showFinishedHint`. Emits: `dismiss-finished`. Styles 1:1 aus `GraphPanel.vue` übernommen, kein Visual-Diff. Sub-Slice 2.1 von 3.
 - **`frontend/src/components/graph/GraphToolbar.vue`** — neue rein präsentationale Header-Toolbar mit Refresh-, GraphML-, SVG-, PNG-, PDF- und Maximize-Buttons. Props: `loading`, `hasGraphId`, `hasGraphData`. Emits: `refresh`, `download-graphml`, `download-svg`, `download-png`, `print-pdf`, `toggle-maximize`. Styles 1:1 aus `GraphPanel.vue` übernommen, kein Visual-Diff. Sub-Slice 2.2 von 3.
+- **`frontend/src/components/graph/GraphCanvas.vue`** — neue Komponente mit der gesamten D3-Renderlogik, SVG-Container, Selektions-State (`selectedItem`, `expandedSelfLoops`), Edge-Labels-Toggle, Loading/Empty-States, Hint- und Detail-Panel-Einbindung sowie allen vier Export-Funktionen (`downloadGraphml`, `downloadSvg`, `downloadPng`, `printPdf`). Exportiert die vier Funktionen via `defineExpose`, sodass die Toolbar-Click-Events vom Composer aus per `canvasRef.value.X()` aufgerufen werden. Props: `graphData` (bereits round-gefiltert), `entityTypes`, `loading`, `currentPhase`, `isSimulating`, `showFinishedHint`. Emits: `dismiss-finished-hint`. Styles 1:1 aus `GraphPanel.vue` übernommen. Sub-Slice 2.3 von 3 — schließt Issue #34 ab.
 
 ### Geändert
 
-- **`frontend/src/components/GraphPanel.vue`** schlanker (933 → 736 Zeilen): Hint- und Toolbar-Markup plus zugehörige Styles in `GraphHints.vue` und `GraphToolbar.vue` ausgelagert. State `showSimulationFinishedHint`/`wasSimulating` plus der `watch` auf `props.isSimulating` bleiben hier (Lifecycle-State gehört nicht in präsentationale Kindkomponenten). Export-Funktionen (`downloadGraphml`, `_buildStandaloneSvg`, `downloadSvg`, `downloadPng`, `printGraphPdf`) bleiben vorerst hier — sie greifen auf den `graphSvg`-DOM-Ref zu und wandern in Sub-Slice 2.3 mit dem `GraphCanvas`.
+- **`frontend/src/components/GraphPanel.vue` ist jetzt reine Kompositionsdatei** (933 → **98 Zeilen**, −90 %). Hält nur noch `selectedRound`/`showSimulationFinishedHint`/`wasSimulating` plus `displayedGraphData`/`entityTypes`/`maxRound` als Computed, dazu den `watch` auf `props.isSimulating`. Komponiert `<GraphToolbar>`, `<GraphCanvas>`, `<GraphLegend>`, `<GraphRoundSlider>`. Toolbar-Click-Events werden via `canvasRef`-Methoden in den Canvas durchgereicht. **Issue #34 (EPIC-04-ST-01) abgeschlossen** — Zielstruktur aus dem Backlog vollständig erreicht.
 - **EPIC-02 (Backend-API-Splitting) als bereits erledigt geschlossen.** Die Inventur zum v0.8.0-Slice-Start zeigt, dass der Split aus `backend/app/api/simulation.py` in zehn fokussierte Module bereits in v0.4.0 vollständig umgesetzt wurde. `simulation.py` ist seither ein 17-zeiliger Compatibility-Shim ohne Routen. 48 Routen liegen unter `simulation_bp` thematisch verteilt (Lifecycle 4, Prepare 2, Run 12, Profiles 16, Interviews 4, History 4, Entities 3, Stream 1, Metrics 2), gemeinsame Helfer in `simulation_common.py`. Issue #29 wird retrospektiv durch `docu/2026-05-01-epic02-api-split-status.md` erfüllt; #30, #31, #32, #33 schließen mit Verweis auf v0.4.0-CHANGELOG-Eintrag und dieses Status-Dokument. Konsequenz: v0.8.0-Backlog reduziert sich von 13 auf 8 echte Issues (EPIC-04 ×3, EPIC-05 ×4, EPIC-10 ×1).
 
 ## [0.7.0] — 2026-05-01
